# lib/api-zod/src/generated/types/rewardAward.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mint Bucks promotional credit system for Mint Printworks
 * OpenAPI spec version: 0.1.0
 */
import type { RewardAwardStatus } from './rewardAwardStatus';

export interface RewardAward {
  id: number;
  ruleId: number;
  /** @nullable */
  ruleName?: string | null;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerCompany?: string | null;
  printavoInvoiceId: string;
  /** @nullable */
  printavoVisualId?: string | null;
  /**
     * Invoice nickname/job name from Printavo, captured when the award was created.
     * @nullable
     */
  nickname?: string | null;
  /**
     * Total order value of the invoice, captured when the award was created.
     * @nullable
     */
  invoiceTotal?: number | null;
  amount: number;
  /**
     * Date (YYYY-MM-DD) the invoice was last paid, captured when the award was created.
     * @nullable
     */
  datePaid?: string | null;
  status: RewardAwardStatus;
  /** @nullable */
  creditId?: number | null;
  /** @nullable */
  note?: string | null;
  awardedAt: string;
  /** @nullable */
  issuedAt?: string | null;
}