# lib/api-zod/src/generated/types/rewardsPipelineItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mint Bucks promotional credit system for Mint Printworks
 * OpenAPI spec version: 0.1.0
 */
import type { RewardsPipelineItemStage } from './rewardsPipelineItemStage';

export interface RewardsPipelineItem {
  printavoInvoiceId: string;
  printavoVisualId: string;
  stage: RewardsPipelineItemStage;
  /**
     * Order nickname (job title) from Printavo.
     * @nullable
     */
  nickname?: string | null;
  /**
     * Date (YYYY-MM-DD) of the most recent payment, null if none.
     * @nullable
     */
  datePaid?: string | null;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerCompany?: string | null;
  customerLinked: boolean;
  /** @nullable */
  total?: number | null;
  /** @nullable */
  amountPaid?: number | null;
  ruleId: number;
  ruleName: string;
  potentialAmount: number;
  createdAt: string;
}